requirements.txt
langchain
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
bs4
httpx
pymupdf
requests
duckduckgo-search
anthropic
lxml